:art: FORMAT (model): move project model to root
